refatorar código após aulas da semana 1

+++ src/Conta.java
@@ -1,4 +1,6 @@
+import java.math.BigDecimal;
+
 public class Conta {
-    public double saldoAtual = 0;
+    public BigDecimal saldoAtual = BigDecimal.valueOf(0);
     public String nome;
 }
